+++ src/frontend/translations/en.ts
@@ -4,12 +4,12 @@
 <context>
     <name>Porn_Fetch</name>
     <message>
-        <location filename="../../../main.py" line="1062"/>
+        <location filename="../../../main.py" line="1070"/>
         <source>Saved User Settings, please restart Porn Fetch!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1099"/>
+        <location filename="../../../main.py" line="1107"/>
         <source>
 FFmpeg isn&apos;t installed on your system... Some features won&apos;t be available:
 
@@ -27,62 +27,62 @@ This warning won&apos;t be shown again.
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1141"/>
+        <location filename="../../../main.py" line="1149"/>
         <source>The output path does not exist or is not writable.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1143"/>
+        <location filename="../../../main.py" line="1151"/>
         <source>Enter custom Path:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1145"/>
+        <location filename="../../../main.py" line="1153"/>
         <source>Success: {text} will be used for this session!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1148"/>
+        <location filename="../../../main.py" line="1156"/>
         <source>Invalid path. The application will now exit.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1472"/>
+        <location filename="../../../main.py" line="1480"/>
         <source>Some error occurred in loading a video. Please report this: {error_message}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1608"/>
+        <location filename="../../../main.py" line="1621"/>
         <source>Information: The Website {website} specified in the URL file isn&apos;t valid.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1626"/>
+        <location filename="../../../main.py" line="1639"/>
         <source>Those credentials don&apos;t seem to be valid...</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1632"/>
+        <location filename="../../../main.py" line="1645"/>
         <source>Login Successful!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1636"/>
+        <location filename="../../../main.py" line="1649"/>
         <source>Login Failed, please check your credentials and try again!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1639"/>
+        <location filename="../../../main.py" line="1652"/>
         <source>You are already logged in!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1662"/>
+        <location filename="../../../main.py" line="1675"/>
         <source>There&apos;s a problem with the login. Please make sure you login first and then you try to get videos based on your account.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../../main.py" line="1732"/>
+        <location filename="../../../main.py" line="1745"/>
         <source>Invalid Category. Press &apos;list categories&apos; to see all possible ones.</source>
         <translation type="unfinished"></translation>
     </message>
@@ -95,74 +95,69 @@ This warning won&apos;t be shown again.
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="772"/>
+        <location filename="../form_desktop.ui" line="782"/>
         <source>PornHub:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="806"/>
+        <location filename="../form_desktop.ui" line="888"/>
         <source>HQPorner:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="878"/>
+        <location filename="../form_desktop.ui" line="875"/>
         <source>Info:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1291"/>
+        <location filename="../form_desktop.ui" line="1305"/>
         <source>Duration (minutes)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1366"/>
+        <location filename="../form_desktop.ui" line="1380"/>
         <source>Only Title (a lot faster)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1458"/>
+        <location filename="../form_desktop.ui" line="1472"/>
         <source>Select everything</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1511"/>
+        <location filename="../form_desktop.ui" line="1525"/>
         <source>Unselect all</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1483"/>
+        <location filename="../form_desktop.ui" line="1497"/>
         <source>Download Selected Videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1286"/>
+        <location filename="../form_desktop.ui" line="1300"/>
         <source>Author</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1318"/>
+        <location filename="../form_desktop.ui" line="1332"/>
         <source>Get Title, Author, Duration</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1386"/>
+        <location filename="../form_desktop.ui" line="1400"/>
         <source>Show videos in reverse</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../form_desktop.ui" line="361"/>
-        <location filename="../form_desktop.ui" line="1135"/>
+        <location filename="../form_desktop.ui" line="1149"/>
         <source>EPorner</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../form_desktop.ui" line="436"/>
         <source>URLs in the file must be separated with new lines!</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../form_desktop.ui" line="653"/>
-        <source>Enter your PornHub Username</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -191,11 +186,6 @@ This warning won&apos;t be shown again.
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="615"/>
-        <source>Username:</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <location filename="../form_desktop.ui" line="701"/>
         <source>Enter your PornHub Password</source>
         <translation type="unfinished"></translation>
@@ -203,10 +193,10 @@ This warning won&apos;t be shown again.
     <message>
         <location filename="../form_desktop.ui" line="275"/>
         <location filename="../form_desktop.ui" line="502"/>
-        <location filename="../form_desktop.ui" line="973"/>
-        <location filename="../form_desktop.ui" line="1002"/>
-        <location filename="../form_desktop.ui" line="1123"/>
-        <location filename="../form_desktop.ui" line="1187"/>
+        <location filename="../form_desktop.ui" line="987"/>
+        <location filename="../form_desktop.ui" line="1016"/>
+        <location filename="../form_desktop.ui" line="1137"/>
+        <location filename="../form_desktop.ui" line="1201"/>
         <source>Get Videos</source>
         <translation type="unfinished"></translation>
     </message>
@@ -241,102 +231,95 @@ This warning won&apos;t be shown again.
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1991"/>
-        <location filename="../form_desktop.ui" line="2042"/>
+        <location filename="../form_desktop.ui" line="2007"/>
+        <location filename="../form_desktop.ui" line="2058"/>
         <source>No</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1984"/>
-        <location filename="../form_desktop.ui" line="2032"/>
+        <location filename="../form_desktop.ui" line="2000"/>
+        <location filename="../form_desktop.ui" line="2048"/>
         <source>Yes</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1737"/>
+        <location filename="../form_desktop.ui" line="1753"/>
         <source>High Performance</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1757"/>
+        <location filename="../form_desktop.ui" line="1773"/>
         <source>Default</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1747"/>
+        <location filename="../form_desktop.ui" line="1763"/>
         <source>FFMPEG</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../form_desktop.ui" line="550"/>
-        <location filename="../form_desktop.ui" line="1586"/>
-        <location filename="../form_desktop.ui" line="1617"/>
-        <location filename="../form_desktop.ui" line="1648"/>
-        <location filename="../form_desktop.ui" line="1679"/>
-        <location filename="../form_desktop.ui" line="1710"/>
-        <location filename="../form_desktop.ui" line="1767"/>
-        <location filename="../form_desktop.ui" line="1893"/>
-        <location filename="../form_desktop.ui" line="1998"/>
-        <location filename="../form_desktop.ui" line="2052"/>
-        <location filename="../form_desktop.ui" line="2104"/>
+        <location filename="../form_desktop.ui" line="1602"/>
+        <location filename="../form_desktop.ui" line="1633"/>
+        <location filename="../form_desktop.ui" line="1664"/>
+        <location filename="../form_desktop.ui" line="1695"/>
+        <location filename="../form_desktop.ui" line="1726"/>
+        <location filename="../form_desktop.ui" line="1783"/>
+        <location filename="../form_desktop.ui" line="1909"/>
+        <location filename="../form_desktop.ui" line="2014"/>
+        <location filename="../form_desktop.ui" line="2068"/>
+        <location filename="../form_desktop.ui" line="2120"/>
         <source>Help</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1943"/>
+        <location filename="../form_desktop.ui" line="1959"/>
         <source>Half</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1933"/>
+        <location filename="../form_desktop.ui" line="1949"/>
         <source>Best</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1953"/>
+        <location filename="../form_desktop.ui" line="1969"/>
         <source>Worst</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2022"/>
+        <location filename="../form_desktop.ui" line="2038"/>
         <source>Use Directory system? </source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1839"/>
+        <location filename="../form_desktop.ui" line="1855"/>
         <source>Enter &quot;./&quot; for current directory</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1832"/>
+        <location filename="../form_desktop.ui" line="1848"/>
         <source>Output path:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2140"/>
+        <location filename="../form_desktop.ui" line="2156"/>
         <source>German</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1037"/>
-        <location filename="../form_desktop.ui" line="1149"/>
+        <location filename="../form_desktop.ui" line="1051"/>
+        <location filename="../form_desktop.ui" line="1163"/>
         <source>Get videos by category</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="871"/>
-        <source>Eporner</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <location filename="../form_desktop.ui" line="368"/>
-        <location filename="../form_desktop.ui" line="786"/>
         <source>XNXX</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../form_desktop.ui" line="354"/>
-        <location filename="../form_desktop.ui" line="779"/>
         <source>XVideos</source>
         <translation type="unfinished"></translation>
     </message>
@@ -372,93 +355,123 @@ This warning won&apos;t be shown again.
     </message>
     <message>
         <location filename="../form_desktop.ui" line="347"/>
-        <location filename="../form_desktop.ui" line="905"/>
+        <location filename="../form_desktop.ui" line="919"/>
         <source>HQPorner</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1024"/>
-        <location filename="../form_desktop.ui" line="1209"/>
+        <location filename="../form_desktop.ui" line="1038"/>
+        <location filename="../form_desktop.ui" line="1223"/>
         <source>List of all categories</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="980"/>
+        <location filename="../form_desktop.ui" line="994"/>
         <source>Get Top Porn:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="954"/>
+        <location filename="../form_desktop.ui" line="968"/>
         <source>Week</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="922"/>
+        <location filename="../form_desktop.ui" line="936"/>
         <source>Month</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="938"/>
+        <location filename="../form_desktop.ui" line="952"/>
         <source>All Time</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1079"/>
+        <location filename="../form_desktop.ui" line="1093"/>
         <source>Get random video</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1101"/>
+        <location filename="../form_desktop.ui" line="1115"/>
         <source>Get Video</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1066"/>
+        <location filename="../form_desktop.ui" line="1080"/>
         <source>Get Brazzers videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1727"/>
+        <location filename="../form_desktop.ui" line="615"/>
+        <source>E-Mail:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="653"/>
+        <source>Enter your PornHub E-Mail address (not your username, pornhub changed it) </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="808"/>
+        <source>XVideos:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="854"/>
+        <source>Eporner:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="861"/>
+        <source>XNXX:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="895"/>
+        <source>Spankbang:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../form_desktop.ui" line="1743"/>
         <source>Threading Mode:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1923"/>
+        <location filename="../form_desktop.ui" line="1939"/>
         <source>Quality:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1977"/>
+        <location filename="../form_desktop.ui" line="1993"/>
         <source>Skip existing files:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2076"/>
+        <location filename="../form_desktop.ui" line="2092"/>
         <source>Model videos (PornHub)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2083"/>
+        <location filename="../form_desktop.ui" line="2099"/>
         <source>User uploads</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2090"/>
+        <location filename="../form_desktop.ui" line="2106"/>
         <source>Featured videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2097"/>
+        <location filename="../form_desktop.ui" line="2113"/>
         <source>Both</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2193"/>
+        <location filename="../form_desktop.ui" line="2209"/>
         <source>French</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2261"/>
+        <location filename="../form_desktop.ui" line="2292"/>
         <source>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
 &lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;meta charset=&quot;utf-8&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
 p, li { white-space: pre-wrap; }
@@ -475,6 +488,7 @@ li.checked::marker { content: &quot;\2612&quot;; }
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;- Eporner.com&lt;/p&gt;
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;- XNXX.com (supports total progress)&lt;/p&gt;
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;- XVideos.com (supports total progress)&lt;/p&gt;
+&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;- Spankbang.com (supports total progress)&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;All sites support *threaded* downloads and selectable quality!&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
@@ -498,7 +512,8 @@ li.checked::marker { content: &quot;\2612&quot;; }
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;- XNXX.com&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
 &lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;I am constantly working to support more websites.&lt;/p&gt;
-&lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
+&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;If you want a specific site to be supported, just ask:&lt;br /&gt;&lt;br /&gt;Discord: echteralsfake&lt;/p&gt;
+&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;GitHub.com/echteralsfake/Porn_Fetch/issues&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;
@@ -508,37 +523,37 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2317"/>
+        <location filename="../form_desktop.ui" line="2350"/>
         <source>Porn Fetch Menu</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2160"/>
+        <location filename="../form_desktop.ui" line="2176"/>
         <source>English</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2183"/>
+        <location filename="../form_desktop.ui" line="2199"/>
         <source>Graphical User Interface Language:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2170"/>
+        <location filename="../form_desktop.ui" line="2186"/>
         <source>System default</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2150"/>
+        <location filename="../form_desktop.ui" line="2166"/>
         <source>Chinese (simplified)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1849"/>
+        <location filename="../form_desktop.ui" line="1865"/>
         <source>Open</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1873"/>
+        <location filename="../form_desktop.ui" line="1889"/>
         <source>Result Limit:</source>
         <translation type="unfinished"></translation>
     </message>
@@ -563,12 +578,12 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1393"/>
+        <location filename="../form_desktop.ui" line="1407"/>
         <source>Export video URLs</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1376"/>
+        <location filename="../form_desktop.ui" line="1390"/>
         <source>Do not clear videos</source>
         <translation type="unfinished"></translation>
     </message>
@@ -578,88 +593,253 @@ li.checked::marker { content: &quot;\2612&quot;; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2129"/>
+        <location filename="../form_desktop.ui" line="2145"/>
         <source>Graphical User Interface</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1824"/>
+        <location filename="../form_desktop.ui" line="1840"/>
         <source>Videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1558"/>
+        <location filename="../form_desktop.ui" line="1574"/>
         <source>Performance</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1505"/>
+        <location filename="../form_desktop.ui" line="1519"/>
         <source>Unselects all videos in the tree widget</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1412"/>
+        <location filename="../form_desktop.ui" line="1426"/>
         <source>Does not stop downloading videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1415"/>
+        <location filename="../form_desktop.ui" line="1429"/>
         <source>Stop loading videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1445"/>
+        <location filename="../form_desktop.ui" line="1459"/>
         <source>Automatically checks a range of videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1451"/>
+        <location filename="../form_desktop.ui" line="1465"/>
         <source>Select a range of videos</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1566"/>
+        <location filename="../form_desktop.ui" line="1582"/>
         <source>Simultaneous downloads:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1597"/>
+        <location filename="../form_desktop.ui" line="1613"/>
         <source>Maximal workers:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1690"/>
+        <location filename="../form_desktop.ui" line="1706"/>
         <source>PornHub Delay (0 = Disabled) in seconds:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1628"/>
+        <location filename="../form_desktop.ui" line="1644"/>
         <source>Maximal timeout:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1659"/>
+        <location filename="../form_desktop.ui" line="1675"/>
         <source>Maximal retries:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1786"/>
+        <location filename="../form_desktop.ui" line="1802"/>
         <source>Apply  (needs restart)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="1815"/>
+        <location filename="../form_desktop.ui" line="1831"/>
         <source>Reset Porn Fetch to default settings</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../form_desktop.ui" line="2205"/>
+        <location filename="../form_desktop.ui" line="2234"/>
         <source>Download and Setup FFmpeg</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../form_desktop.ui" line="193"/>
         <source>Start</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>main</name>
+    <message>
+        <location filename="../../../main.py" line="67"/>
+        <source>Wrong Input, please verify the URL, category or actress!</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../../main.py" line="193"/>
+        <source>Disabled</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../../main.py" line="1880"/>
+        <source>Done! Please restart.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../../main.py" line="1904"/>
+        <source>No URLs in the current session...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../../main.py" line="1907"/>
+        <source>FFmpeg has been installed. Please restart Porn Fetch :)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../../main.py" line="1929"/>
+        <source>
+            Information: A new version of Porn Fetch (v{__next_release__}) is out. I recommend you to update Porn Fetch. 
+            Go to: https://github.com/EchterAlsFake/Porn_Fetch/releases/tag/ {__next_release__}
+
+            Changelog:
+            {markdown.markdown(changelog)}
+
+            </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="11"/>
+        <source>
+The result limit defines how many videos will be returned when performing a search or doing other operations which
+involves loading multiple videos. This also affects models / channels and your liked videos. The result limit is
+basically the number of videos which can be loaded into the tree widget (this thing where videos are displayed).
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="24"/>
+        <source>
+You can set a delay between requests from you to PornHub. If you are downloading a lot of videos or experiencing 
+&apos;client.call&apos; errors, you should enable a delay. By default the delay is turned off with the value 0
+
+A good starting point is between 0.5 - 1.5
+
+The longer the delay is, the longer it will take to download videos, load videos and generally do stuff.
+This does NOT affect other sites!
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="36"/>
+        <source>
+The maximal workers define the amount of maximal threads which can be started when using the threaded download mode.
+One thread handles downloading one segment, so (in theory) 20 threads can download 20 segments at the same time.
+This can of course be helpful when you have a very fast internet connection, but when you have a poor PC or running on
+Android, you should set this to a lower value.
+
+I recommend &apos;3&apos; for Android and 5 for low bandwidth connections &lt; 15000 bit/s
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="44"/>
+        <source>
+The timeout handles the timeout for retrieving segments when using the threaded download mode. If you have a poor 
+internet connection you can set this higher than 10. But this isn&apos;t required for most users!
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="55"/>
+        <source>
+The Semaphore is a tool to limit the number of simultaneous actions / downloads.
+
+For example: If the semaphore is set to 1, only 1 video will be downloaded at the same time.
+If the semaphore is set to 4, 4 videos will be downloaded at the same time. Changing this is only useful, if
+you have a really good internet connection and a good system.
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="73"/>
+        <source>
+The different threading modes are used for different scenarios. 
+
+1) High Performance:  Uses a class of workers to download multiple video segments at a time. Can be really fast if you
+have a very strong internet connection. Maybe not great for low end systems.
+
+2) FFMPEG:  ffmpeg is a tool for converting media files. ffmpeg will download every video segment and merge it directly
+into the video file. This removes an extra step from the default method and is therefore a lot faster, but still not as 
+good as high performance.
+
+3) Default:  The default download mode will just download one video segment after the next one. If you get a lot of 
+timeouts this can really slow down the process, as we need to wait for PornHub to return the video segments.
+With the High Performance method, we can just download other segments while waiting which makes it so fast.
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="84"/>
+        <source>
+The directory system will save videos in an intelligent way. If you download 3 videos form one Pornstar and 5 videos 
+from another, Porn Fetch will automatically make folders for it and move the 3 videos into that one folder and the other
+5 into the other. (This will still apply with your selected output path)
+
+This can be helpful for organizing stuff, but is a more advanced feature, so the majority of users won&apos;t use it probably
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="107"/>
+        <source>
+Create a .txt file and add URLs like this:
+
+url1
+url2
+url3
+...
+
+Split them with new lines. No comma, not multiple URLs in the same line!
+You can also add model URLs like this:
+
+model#MODEL_URL
+
+An example for a file would be:
+
+https://de.pornhub.com/view_video.php?viewkey=ph5be76343323ff
+https://de.pornhub.com/view_video.php?viewkey=ph5946e5f19585a
+model#https://de.pornhub.com/pornstar/nancy-a
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="115"/>
+        <source>
+The maximal retries defines how much attempts will be used for a network request. For example if an API calls
+a URL for a website there will be &lt;AMOUNT&gt; of attempts until an error is thrown.
+</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../backend/class_help.py" line="143"/>
+        <source>
+User uploads and featured videos are two different things. User uploads are the videos which were really uploaded
+by the model and the featured videos are videos the model is part or featured in.
+
+For example the model Nancy Ace has like 10 self uploaded which she made by herself, but she is part in like thousands
+of videos from other studios.
+
+If you choose &quot;User Uploads&quot;, only self uploaded videos will be fetched, and the other way around :)</source>
         <translation type="unfinished"></translation>
     </message>
 </context>